moves all logic for matches into MatchingModal, moves MatchingModal into UserManagement.js, fixes refresh bug, and adds loading component
