--- src/components/pages/MentorMenteeMatching/matches.json
@@ -1,44 +1,0 @@
-[
-  {
-    "mentee_ids": ["ac985dd8-9164-4825-833b-aef12264db8f"],
-    "mentor_id": "601d51ed-6016-4fa9-ae49-fe50b082e8c3",
-    "created_at": "2022-10-31T19:36:48.231000",
-    "status": "mentee"
-  },
-  {
-    "mentee_ids": ["ac985dd8-9164-4825-833b-aef12264db8f"],
-    "mentor_id": "fce3b89c-4d8e-4547-bdfc-2ffa6e7b1492",
-    "created_at": "2022-11-2T18:22:38:581000",
-    "status": "mentee"
-  },
-  {
-    "mentee_ids": ["c6af50a3-797a-4e09-b3b2-2cbf0b2baea3"],
-    "mentor_id": "601d51ed-6016-4fa9-ae49-fe50b082e8c3",
-    "created_at": "2022-11-2T18:22:38:581000",
-    "status": "mentee"
-  },
-
-  {
-    "mentee_ids": ["c6af50a3-797a-4e09-b3b2-2cbf0b2baea3"],
-    "mentor_id": "57c73735-def9-4d2d-9587-617e246422e8",
-    "created_at": "2022-11-2T18:22:38:581000",
-    "status": "mentee"
-  },
-
-  {
-    "mentee_ids": ["c6af50a3-797a-4e09-b3b2-2cbf0b2baea3"],
-    "mentor_id": "7e8b7ff4-eb28-4688-b7ee-4c5008db11c7",
-    "created_at": "2022-11-2T18:22:38:581000",
-    "status": "mentee"
-  },
-
-  {
-    "mentee_ids": [
-      "c6af50a3-797a-4e09-b3b2-2cbf0b2baea3",
-      "f6ebc220-bd97-4ef7-a364-8d4a016fe38a"
-    ],
-    "mentor_id": "601d51ed-6016-4fa9-ae49-fe50b082e8c3",
-    "created_at": "2022-11-2T18:22:38:581000",
-    "status": "mentee"
-  }
-]
